perf(feature-flags): use a Set for isValidFeatureId in a public utils module
Avoid a per-call FeatureIdSchema.safeParse() allocation in the type guard by
precomputing a Set for O(1) lookups in hot paths (e.g. app filtering). Move the
guard out of schema.ts into a dedicated public utils.ts, and rename the internal
resolution helpers (resolveFeature, resolveAll, ...) to internals.ts (kept out of
the package barrel).
