+++ shared/feature-flags/src/data/index.ts
@@ -1,4 +1,5 @@
 export * from "./schema";
+export * from "./utils";
 export * from "./slice";
 export * from "./selectors";
 export * from "./middleware";
